+++ catalog/app/components/Preview/loaders/Csv.js
@@ -3,9 +3,11 @@ import * as R from 'ramda'
 import { PreviewData } from '../types'
 import * as utils from './utils'
 
-export const detect = R.pipe(utils.stripCompression, utils.extIn(['.csv', '.tsv']))
+export const isCsv = R.pipe(utils.stripCompression, utils.extIs('.csv'))
 
-const isTsv = R.pipe(utils.stripCompression, utils.extIs('.tsv'))
+export const isTsv = R.pipe(utils.stripCompression, utils.extIs('.tsv'))
+
+export const detect = R.anyPass([isCsv, isTsv])
 
 export const Loader = function CsvLoader({ handle, children }) {
   const data = utils.usePreview({