+++ catalog/app/components/Preview/types.js
@@ -24,6 +24,7 @@ PreviewStatus: {
 */
 
 export const PreviewData = tagged([
+  'Audio', // { src: string }
   'DataFrame', // { preview: string, ...PreviewStatus }
   'ECharts', // { option: object }
   'Fcs', // { preview: string, metadata: object, ...PreviewStatus }
@@ -34,11 +35,11 @@ export const PreviewData = tagged([
   'Notebook', // { preview: string, ...PreviewStatus }
   'Parquet', // { preview: string, ...ParquetMeta, ...PreviewStatus }
   'Pdf', // { handle: object, pages: number, firstPageBlob: Blob, type: 'pdf' | 'pptx' }
+  'Perspective', // { context: CONTEXT, data: string | ArrayBuffer, handle: S3Handle, onLoadMore: () => void, truncated: boolean }
   'Text', // { head: string, tail: string, lang: string, highlighted: { head: string, tail: string }, ...PreviewStatus }
   'Vcf', // { meta: string[], header: string[], body: string[][], variants: string[], ...PreviewStatus }
   'Vega', // { spec: Object }
   'Video', // { src: string }
-  'Audio', // { src: string }
   'Voila', // { src: string }
 ])
 
@@ -55,3 +56,8 @@ export const PreviewError = tagged([
   'MalformedJson', // { handle, message }
   'Unexpected', // { handle, retry, originalError: any }
 ])
+
+export const CONTEXT = {
+  FILE: 'file',
+  LISTING: 'listing',
+}