+++ catalog/app/containers/Bucket/PackageTree/PackageTree.tsx
@@ -434,7 +434,9 @@ const withPreview = (
   if (archived) {
     return callback(AsyncResult.Err(Preview.PreviewError.Archived({ handle })))
   }
-  return Preview.load({ ...handle, mode, packageHandle }, callback)
+  const previewHandle = { ...handle, packageHandle }
+  const previewOptions = { mode, context: Preview.CONTEXT.LISTING }
+  return Preview.load(previewHandle, callback, previewOptions)
 }
 
 interface ObjectAttrs {