+++ catalog/app/containers/Bucket/Summarize.tsx
@@ -280,11 +280,22 @@ export function FilePreview({
   const key = handle.logicalKey || handle.key
   const props = React.useMemo(() => Preview.getRenderProps(key, file), [key, file])
 
+  const previewOptions = React.useMemo(
+    () => ({
+      ...file,
+      context: Preview.CONTEXT.LISTING,
+    }),
+    [file],
+  )
+  const previewHandle = React.useMemo(
+    () => ({ ...handle, packageHandle }),
+    [handle, packageHandle],
+  )
   // TODO: check for glacier and hide items
   return (
     <Section description={description} heading={heading} handle={handle}>
       {Preview.load(
-        { ...handle, packageHandle }, // FIXME: memoize
+        previewHandle,
         Preview.display({
           renderContents: (contents: React.ReactNode) => (
             <PreviewBox {...{ contents, expanded }} />
@@ -292,7 +303,7 @@ export function FilePreview({
           renderProgress: () => <ContentSkel />,
           props,
         }),
-        file,
+        previewOptions,
       )}
     </Section>
   )