+++ catalog/app/embed/File.js
@@ -346,6 +346,8 @@ const useStyles = M.makeStyles((t) => ({
   },
 }))
 
+const previewOptions = { context: Preview.CONTEXT.FILE }
+
 export default function File({
   match: {
     params: { bucket, path: encodedPath },
@@ -420,7 +422,7 @@ export default function File({
         if (h.archived) {
           return callback(AsyncResult.Err(Preview.PreviewError.Archived({ handle })))
         }
-        return Preview.load(handle, callback)
+        return Preview.load(handle, callback, previewOptions)
       },
       DoesNotExist: () =>
         callback(AsyncResult.Err(Preview.PreviewError.InvalidVersion({ handle }))),