+++ catalog/app/utils/perspective-pollution.ts
@@ -1,0 +1,7 @@
+/* eslint-disable import/no-unresolved */
+//   Wait till eslint will handle `exports` field
+//   https://github.com/import-js/eslint-plugin-import/issues/1810
+import '@finos/perspective-viewer'
+import '@finos/perspective-viewer-datagrid'
+import '@finos/perspective-viewer-d3fc'
+import '@finos/perspective-viewer/dist/css/material.css'